separate inference related conversions
