+++ lib/api/src/conversions/mod.rs
@@ -1,1 +1,2 @@
+pub mod inference;
 pub mod vectors;
